\name{guitar}
\alias{guitar}
\title{
D-string of acoustic guitar
}
\description{
Time series of a an oscillating nylon D-string of a 3/4 acoustic guitar. 
}
\usage{
guitar
}

\format{
A data frame with variables \code{time} in seconds and \code{volume} in decibel of an oscillating D4-string, which corresponds to a frequency of around 293. The volume is recorded by a comparatively bad microphone.
}

\keyword{datasets}
