\encoding{utf8}
\name{robts-package}
\alias{robts-package}
\docType{package}
\title{
Robust Time Series Analysis
}
\description{
Collection of R functions for robust time series analysis.  This includes various approaches for robust estimation of (partial) autocorrelation, robust filtering, robust fitting of AR(p) processes as well as robust changepoint detection.
}
\details{
See the main functions \code{\link{acfrob}} for robust estimation of the autocorrelation function, \code{\link{arrob}} for robust fitting of autoregressive (AR) models, \code{link{robspectrum}} for robust estimation of the spectral density and \code{\link{rob.change}} for robust change point tests.
}
\author{
Alexander \enc{Dürre}{Duerre}, Tobias Liboschik, Roland Fried and Jonathan Rathjens

Maintainer: Alexander \enc{Dürre}{Duerre} <alexander.duerre@udo.edu>
}
\references{
\enc{Dürre}{Duerre}, A., Fried, R. and Liboschik, T. (2015): Robust estimation of (partial) autocorrelation, \emph{Wiley Interdisciplinary Reviews: Computational Statistics}, vol. 7, 205--222.

Maronna, R. A., Martin, R. D., and Yohai, V. J. (2006): \emph{Robust Statistics: Theory and Methods}, Wiley, chapter 8.
}
