\name{sunhours}
\alias{sunhours}

\title{
Sun Hours Per Year in a Location in Chemnitz 
}
\description{
Yearly sun hours measured in Chemnitz, Neukirchener \enc{Straße}{Strasse} (Germany). Recordings started in 1951.
}
\usage{
sunhours
}

\format{
A data frame with variables \code{year} and \code{sunhours} giving the total number of sun hours per year. 
}

\source{
Deutscher Wetterdienst: SurvStat@RKI, \url{ftp://ftp-cdc.
dwd.de/pub/CDC/observations_germany/climate/}, accessed on 15th December 2015.
}

\keyword{Data}
