\encoding{utf8}
\name{robts-package}
\alias{robts-package}
\docType{package}
\title{
Robust Time Series Analysis
}
\description{
Collection of R functions for robust time series analysis.  This includes various approaches for robust estimation of (partial) autocorrelation, robust filtering, robust fitting of AR(p) processes and robust changepoint detection.
}
\details{
\tabular{ll}{
Package: \tab robts\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-11-11\cr
License: \tab GPL-2 | GPL-3\cr
}
}
\author{
Alexander \enc{Dürre}{Duerre}, Tobias Liboschik, Roland Fried and Jonathan Rathjens

Maintainer: Jonathan Rathjens <jonathan.rathjens@tu-dortmund.de>
}
\references{
\enc{Dürre}{Duerre}, A., Fried, R. and Liboschik, T. (2014): Robust estimation of (partial) autocorrelation, working paper, \url{http://hdl.handle.net/2003/33011}.

Maronna, R. A., Martin, R. D., and Yohai, V. J. (2006): \emph{Robust Statistics: Theory and Methods}, Wiley, chapter 8.
}
